@@ -79,15 +79,15 @@
         "@swc/core": "^1.3.107",
         "@swc/jest": "^0.2.36",
         "@testing-library/jest-dom": "^6.4.2",
-        "@testing-library/react": "12.1.2",
+        "@testing-library/react": "^15.0.5",
         "@types/chrome": "0.0.260",
         "@types/jest": "^29.5.11",
         "@types/jsdom": "^21.1.6",
         "@types/lodash": "^4.14.202",
         "@types/luxon": "^3.4.2",
         "@types/node": "^16.11.7",
-        "@types/react": "^16.14.25",
-        "@types/react-dom": "^16.9.15",
+        "@types/react": "^18.3.1",
+        "@types/react-dom": "^18.3.0",
         "@types/react-helmet": "^6.1.11",
         "@types/react-router-dom": "^5.3.3",
         "@types/serve-static": "^1.15.5",
@@ -129,7 +129,7 @@
         "license-check-and-add": "^4.0.5",
         "mini-css-extract-plugin": "2.9.0",
         "npm-run-all": "^4.1.5",
-        "playwright": "^1.44.0",
+        "playwright": "^1.44.1",
         "postcss": "^8.4.35",
         "postcss-modules": "^6.0.0",
         "prettier": "^3.2.5",
@@ -151,8 +151,8 @@
         "webpack-node-externals": "^3.0.0"
     },
     "dependencies": {
-        "@fluentui/react": "^8.96.1",
         "@fluentui/react-components": "^9.53.0",
+        "@fluentui/react": "^8.118.1",
         "@microsoft/applicationinsights-web": "^2.8.15",
         "@testing-library/user-event": "^14.5.2",
         "ajv": "^8.12.0",
@@ -161,9 +161,9 @@
         "idb-keyval": "^6.2.1",
         "lodash": "^4.17.21",
         "luxon": "^3.4.4",
-        "react": "^16.14.0",
-        "react-dom": "^16.14.0",
-        "react-helmet": "^6.1.0",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "react-helmet-async": "^2.0.5",
         "react-resize-detector": "^9.1.1",
         "react-router-dom": "^6.21.3",
         "tabbable": "^6.2.0",
